Routing adjustments
Modified the method access property for the `setModule` and `setAction`
methods. We have to be able to set the module and action from outside the
class if no matching route have been located.

Removed the forced formatting of the action since the formatting will
prevent actions to be camelcased, e.g. NotFound.