Added namespace related functionality
Support to enable/disable namespace usage have been implemented. This usage can
be enabled/disabled on an application or module level.

Implemented better support for namespace fallback. If the module namespace is
not defined but the core namespace is defined, and namespage usage is enabled.
The module namespace will be the core namespace suffixed with the module name.

Removed the module configuration files since they only contained the namespace.
And, with the namespace changes they are not necessary anymore.

Also, added documentation and todo notes.